factor out Diffraction1D
