+++ pyxem/__init__.py
@@ -37,9 +37,12 @@ from .generators.library_generator import DiffractionLibraryGenerator
 from .generators.library_generator import VectorLibraryGenerator
 from .generators.calibration_generator import CalibrationGenerator
 
-from .signals.crystallographic_map import CrystallographicMap
+from .signals.diffraction1D import Diffraction1D
+from .signals.diffraction2D import Diffraction2D
 from .signals.diffraction_profile import ElectronDiffraction1D
 from .signals.electron_diffraction import ElectronDiffraction2D
+
+from .signals.crystallographic_map import CrystallographicMap
 from .signals.diffraction_simulation import DiffractionSimulation
 from .signals.diffraction_vectors import DiffractionVectors
 from .signals.indexation_results import TemplateMatchingResults
