+++ pyxem/signals/electron_diffraction.py
@@ -26,7 +26,7 @@ from pyxem.signals.diffraction2d import Diffraction2D
 
 
 class ElectronDiffraction2D(Diffraction2D):
-    _signal_type = "electron_diffraction"
+    _signal_type = "electron_diffraction2d"
 
     def __init__(self, *args, **kwargs):
         """
